+++ Godeps/_workspace/src/github.com/gin-gonic/gin/binding/binding.go
@@ -4,11 +4,7 @@
 
 package binding
 
-import (
-	"net/http"
-
-	"gopkg.in/bluesuncorp/validator.v5"
-)
+import "net/http"
 
 const (
 	MIMEJSON              = "application/json"
@@ -25,7 +21,16 @@ type Binding interface {
 	Bind(*http.Request, interface{}) error
 }
 
-var validate = validator.New("binding", validator.BakedInValidators)
+type StructValidator interface {
+	// ValidateStruct can receive any kind of type and it should never panic, even if the configuration is not right.
+	// If the received type is not a struct, any validation should be skipped and nil must be returned.
+	// If the received type is a struct or pointer to a struct, the validation should be performed.
+	// If the struct is not valid or the validation itself fails, a descriptive error should be returned.
+	// Otherwise nil must be returned.
+	ValidateStruct(interface{}) error
+}
+
+var Validator StructValidator = &defaultValidator{}
 
 var (
 	JSON = jsonBinding{}
@@ -42,22 +47,15 @@ func Default(method, contentType string) Binding {
 			return JSON
 		case MIMEXML, MIMEXML2:
 			return XML
-		default:
+		default: //case MIMEPOSTForm, MIMEMultipartPOSTForm:
 			return Form
 		}
 	}
 }
 
-func ValidateField(f interface{}, tag string) error {
-	if err := validate.Field(f, tag); err != nil {
-		return error(err)
+func validate(obj interface{}) error {
+	if Validator == nil {
+		return nil
 	}
-	return nil
-}
-
-func Validate(obj interface{}) error {
-	if err := validate.Struct(obj); err != nil {
-		return error(err)
-	}
-	return nil
+	return Validator.ValidateStruct(obj)
 }