+++ Godeps/_workspace/src/github.com/gin-gonic/gin/context_test.go
@@ -8,12 +8,12 @@ import (
 	"bytes"
 	"errors"
 	"html/template"
+	"mime/multipart"
 	"net/http"
 	"net/http/httptest"
 	"testing"
 	"time"
 
-	"github.com/gin-gonic/gin/binding"
 	"github.com/manucorporat/sse"
 	"github.com/stretchr/testify/assert"
 )
@@ -31,6 +31,27 @@ func createTestContext() (c *Context, w *httptest.ResponseRecorder, r *Engine) {
 	c.reset()
 	c.writermem.reset(w)
 	return
+}
+
+func createMultipartRequest() *http.Request {
+	boundary := "--testboundary"
+	body := new(bytes.Buffer)
+	mw := multipart.NewWriter(body)
+	defer mw.Close()
+
+	must(mw.SetBoundary(boundary))
+	must(mw.WriteField("foo", "bar"))
+	must(mw.WriteField("bar", "foo"))
+	req, err := http.NewRequest("POST", "/", body)
+	must(err)
+	req.Header.Set("Content-Type", MIMEMultipartPOSTForm+"; boundary="+boundary)
+	return req
+}
+
+func must(err error) {
+	if err != nil {
+		panic(err.Error())
+	}
 }
 
 func TestContextReset(t *testing.T) {
@@ -52,7 +73,7 @@ func TestContextReset(t *testing.T) {
 	assert.Empty(t, c.Errors.Errors())
 	assert.Empty(t, c.Errors.ByType(ErrorTypeAny))
 	assert.Len(t, c.Params, 0)
-	assert.Equal(t, c.index, -1)
+	assert.EqualValues(t, c.index, -1)
 	assert.Equal(t, c.Writer.(*responseWriter), &c.writermem)
 }
 
@@ -114,46 +135,77 @@ func TestContextCopy(t *testing.T) {
 	assert.Equal(t, cp.Params, c.Params)
 }
 
-func TestContextFormParse(t *testing.T) {
+func TestContextQuery(t *testing.T) {
 	c, _, _ := createTestContext()
 	c.Request, _ = http.NewRequest("GET", "http://example.com/?foo=bar&page=10", nil)
 
-	assert.Equal(t, c.DefaultFormValue("foo", "none"), "bar")
-	assert.Equal(t, c.FormValue("foo"), "bar")
-	assert.Empty(t, c.PostFormValue("foo"))
+	assert.Equal(t, c.DefaultQuery("foo", "none"), "bar")
+	assert.Equal(t, c.Query("foo"), "bar")
+	assert.Empty(t, c.PostForm("foo"))
 
-	assert.Equal(t, c.DefaultFormValue("page", "0"), "10")
-	assert.Equal(t, c.FormValue("page"), "10")
-	assert.Empty(t, c.PostFormValue("page"))
+	assert.Equal(t, c.DefaultQuery("page", "0"), "10")
+	assert.Equal(t, c.Query("page"), "10")
+	assert.Empty(t, c.PostForm("page"))
 
-	assert.Equal(t, c.DefaultFormValue("NoKey", "nada"), "nada")
-	assert.Empty(t, c.FormValue("NoKey"))
-	assert.Empty(t, c.PostFormValue("NoKey"))
+	assert.Equal(t, c.DefaultQuery("NoKey", "nada"), "nada")
+	assert.Empty(t, c.Query("NoKey"))
+	assert.Empty(t, c.PostForm("NoKey"))
 }
 
-func TestContextPostFormParse(t *testing.T) {
+func TestContextQueryAndPostForm(t *testing.T) {
 	c, _, _ := createTestContext()
 	body := bytes.NewBufferString("foo=bar&page=11&both=POST")
 	c.Request, _ = http.NewRequest("POST", "/?both=GET&id=main", body)
 	c.Request.Header.Add("Content-Type", MIMEPOSTForm)
 
-	assert.Equal(t, c.DefaultPostFormValue("foo", "none"), "bar")
-	assert.Equal(t, c.PostFormValue("foo"), "bar")
-	assert.Equal(t, c.FormValue("foo"), "bar")
+	assert.Equal(t, c.DefaultPostForm("foo", "none"), "bar")
+	assert.Equal(t, c.PostForm("foo"), "bar")
+	assert.Empty(t, c.Query("foo"))
 
-	assert.Equal(t, c.DefaultPostFormValue("page", "0"), "11")
-	assert.Equal(t, c.PostFormValue("page"), "11")
-	assert.Equal(t, c.FormValue("page"), "11")
+	assert.Equal(t, c.DefaultPostForm("page", "0"), "11")
+	assert.Equal(t, c.PostForm("page"), "11")
+	assert.Equal(t, c.Query("page"), "")
 
-	assert.Equal(t, c.PostFormValue("both"), "POST")
-	assert.Equal(t, c.FormValue("both"), "POST")
+	assert.Equal(t, c.PostForm("both"), "POST")
+	assert.Equal(t, c.Query("both"), "GET")
 
-	assert.Equal(t, c.FormValue("id"), "main")
-	assert.Empty(t, c.PostFormValue("id"))
+	assert.Equal(t, c.DefaultPostForm("id", "000"), "000")
+	assert.Equal(t, c.Query("id"), "main")
+	assert.Empty(t, c.PostForm("id"))
 
-	assert.Equal(t, c.DefaultPostFormValue("NoKey", "nada"), "nada")
-	assert.Empty(t, c.PostFormValue("NoKey"))
-	assert.Empty(t, c.FormValue("NoKey"))
+	assert.Equal(t, c.DefaultPostForm("NoKey", "nada"), "nada")
+	assert.Empty(t, c.PostForm("NoKey"))
+	assert.Empty(t, c.Query("NoKey"))
+
+	var obj struct {
+		Foo  string `form:"foo"`
+		Id   string `form:"id"`
+		Page string `form:"page"`
+		Both string `form:"both"`
+	}
+	assert.NoError(t, c.Bind(&obj))
+	assert.Equal(t, obj.Foo, "bar")
+	assert.Equal(t, obj.Id, "main")
+	assert.Equal(t, obj.Page, "11")
+	assert.Equal(t, obj.Both, "POST")
+}
+
+func TestContextPostFormMultipart(t *testing.T) {
+	c, _, _ := createTestContext()
+	c.Request = createMultipartRequest()
+
+	var obj struct {
+		Foo string `form:"foo"`
+		Bar string `form:"bar"`
+	}
+	assert.NoError(t, c.Bind(&obj))
+	assert.Equal(t, obj.Bar, "foo")
+	assert.Equal(t, obj.Foo, "bar")
+
+	assert.Empty(t, c.Query("foo"))
+	assert.Empty(t, c.Query("bar"))
+	assert.Equal(t, c.PostForm("foo"), "bar")
+	assert.Equal(t, c.PostForm("bar"), "foo")
 }
 
 // Tests that the response is serialized as JSON
@@ -165,6 +217,18 @@ func TestContextRenderJSON(t *testing.T) {
 	assert.Equal(t, w.Code, 201)
 	assert.Equal(t, w.Body.String(), "{\"foo\":\"bar\"}\n")
 	assert.Equal(t, w.HeaderMap.Get("Content-Type"), "application/json; charset=utf-8")
+}
+
+// Tests that the response is serialized as JSON
+// we change the content-type before
+func TestContextRenderAPIJSON(t *testing.T) {
+	c, w, _ := createTestContext()
+	c.Header("Content-Type", "application/vnd.api+json")
+	c.JSON(201, H{"foo": "bar"})
+
+	assert.Equal(t, w.Code, 201)
+	assert.Equal(t, w.Body.String(), "{\"foo\":\"bar\"}\n")
+	assert.Equal(t, w.HeaderMap.Get("Content-Type"), "application/vnd.api+json")
 }
 
 // Tests that the response is serialized as JSON
@@ -401,13 +465,18 @@ func TestContextClientIP(t *testing.T) {
 	c, _, _ := createTestContext()
 	c.Request, _ = http.NewRequest("POST", "/", nil)
 
-	c.Request.Header.Set("X-Real-IP", "10.10.10.10")
-	c.Request.Header.Set("X-Forwarded-For", "20.20.20.20 , 30.30.30.30")
-	c.Request.RemoteAddr = "40.40.40.40"
+	c.Request.Header.Set("X-Real-IP", " 10.10.10.10  ")
+	c.Request.Header.Set("X-Forwarded-For", "  20.20.20.20, 30.30.30.30")
+	c.Request.RemoteAddr = "  40.40.40.40 "
 
 	assert.Equal(t, c.ClientIP(), "10.10.10.10")
+
 	c.Request.Header.Del("X-Real-IP")
 	assert.Equal(t, c.ClientIP(), "20.20.20.20")
+
+	c.Request.Header.Set("X-Forwarded-For", "30.30.30.30  ")
+	assert.Equal(t, c.ClientIP(), "30.30.30.30")
+
 	c.Request.Header.Del("X-Forwarded-For")
 	assert.Equal(t, c.ClientIP(), "40.40.40.40")
 }
@@ -420,15 +489,31 @@ func TestContextContentType(t *testing.T) {
 	assert.Equal(t, c.ContentType(), "application/json")
 }
 
-func TestContextAutoBind(t *testing.T) {
-	c, w, _ := createTestContext()
+func TestContextAutoBindJSON(t *testing.T) {
+	c, _, _ := createTestContext()
 	c.Request, _ = http.NewRequest("POST", "/", bytes.NewBufferString("{\"foo\":\"bar\", \"bar\":\"foo\"}"))
 	c.Request.Header.Add("Content-Type", MIMEJSON)
+
 	var obj struct {
 		Foo string `json:"foo"`
 		Bar string `json:"bar"`
 	}
 	assert.NoError(t, c.Bind(&obj))
+	assert.Equal(t, obj.Bar, "foo")
+	assert.Equal(t, obj.Foo, "bar")
+	assert.Empty(t, c.Errors)
+}
+
+func TestContextBindWithJSON(t *testing.T) {
+	c, w, _ := createTestContext()
+	c.Request, _ = http.NewRequest("POST", "/", bytes.NewBufferString("{\"foo\":\"bar\", \"bar\":\"foo\"}"))
+	c.Request.Header.Add("Content-Type", MIMEXML) // set fake content-type
+
+	var obj struct {
+		Foo string `json:"foo"`
+		Bar string `json:"bar"`
+	}
+	assert.NoError(t, c.BindJSON(&obj))
 	assert.Equal(t, obj.Bar, "foo")
 	assert.Equal(t, obj.Foo, "bar")
 	assert.Equal(t, w.Body.Len(), 0)
@@ -451,20 +536,6 @@ func TestContextBadAutoBind(t *testing.T) {
 	assert.Empty(t, obj.Foo)
 	assert.Equal(t, w.Code, 400)
 	assert.True(t, c.IsAborted())
-}
-
-func TestContextBindWith(t *testing.T) {
-	c, w, _ := createTestContext()
-	c.Request, _ = http.NewRequest("POST", "/", bytes.NewBufferString("{\"foo\":\"bar\", \"bar\":\"foo\"}"))
-	c.Request.Header.Add("Content-Type", MIMEXML)
-	var obj struct {
-		Foo string `json:"foo"`
-		Bar string `json:"bar"`
-	}
-	assert.NoError(t, c.BindWith(&obj, binding.JSON))
-	assert.Equal(t, obj.Bar, "foo")
-	assert.Equal(t, obj.Foo, "bar")
-	assert.Equal(t, w.Body.Len(), 0)
 }
 
 func TestContextGolangContext(t *testing.T) {