+++ Godeps/_workspace/src/github.com/gin-gonic/gin/debug.go
@@ -4,13 +4,11 @@
 
 package gin
 
-import (
-	"log"
-	"os"
-)
+import "log"
 
-var debugLogger = log.New(os.Stdout, "[GIN-debug] ", 0)
-
+func init() {
+	log.SetFlags(0)
+}
 func IsDebugging() bool {
 	return ginMode == debugCode
 }
@@ -18,19 +16,23 @@ func IsDebugging() bool {
 func debugPrintRoute(httpMethod, absolutePath string, handlers HandlersChain) {
 	if IsDebugging() {
 		nuHandlers := len(handlers)
-		handlerName := nameOfFunction(handlers[nuHandlers-1])
+		handlerName := nameOfFunction(handlers.Last())
 		debugPrint("%-5s %-25s --> %s (%d handlers)\n", httpMethod, absolutePath, handlerName, nuHandlers)
 	}
 }
 
 func debugPrint(format string, values ...interface{}) {
 	if IsDebugging() {
-		debugLogger.Printf(format, values...)
+		log.Printf("[GIN-debug] "+format, values...)
 	}
 }
 
 func debugPrintWARNING() {
-	debugPrint("[WARNING] Running in DEBUG mode! Disable it before going production\n")
+	debugPrint(`[WARNING] Running in "debug" mode. Switch to "release" mode in production.
+ - using env:	export GIN_MODE=release
+ - using code:	gin.SetMode(gin.ReleaseMode)
+
+`)
 }
 
 func debugPrintError(err error) {