+++ Godeps/_workspace/src/github.com/gin-gonic/gin/errors.go
@@ -66,27 +66,40 @@ func (msg *Error) JSON() interface{} {
 	return json
 }
 
+// Implements the json.Marshaller interface
 func (msg *Error) MarshalJSON() ([]byte, error) {
 	return json.Marshal(msg.JSON())
 }
 
+// Implements the error interface
 func (msg *Error) Error() string {
 	return msg.Err.Error()
 }
 
+func (msg *Error) IsType(flags ErrorType) bool {
+	return (msg.Type & flags) > 0
+}
+
+// Returns a readonly copy filterd the byte.
+// ie ByType(gin.ErrorTypePublic) returns a slice of errors with type=ErrorTypePublic
 func (a errorMsgs) ByType(typ ErrorType) errorMsgs {
 	if len(a) == 0 {
+		return nil
+	}
+	if typ == ErrorTypeAny {
 		return a
 	}
-	result := make(errorMsgs, 0, len(a))
+	var result errorMsgs = nil
 	for _, msg := range a {
-		if msg.Type&typ > 0 {
+		if msg.IsType(typ) {
 			result = append(result, msg)
 		}
 	}
 	return result
 }
 
+// Returns the last error in the slice. It returns nil if the array is empty.
+// Shortcut for errors[len(errors)-1]
 func (a errorMsgs) Last() *Error {
 	length := len(a)
 	if length == 0 {
@@ -95,9 +108,17 @@ func (a errorMsgs) Last() *Error {
 	return a[length-1]
 }
 
+// Returns an array will all the error messages.
+// Example
+// ```
+// c.Error(errors.New("first"))
+// c.Error(errors.New("second"))
+// c.Error(errors.New("third"))
+// c.Errors.Errors() // == []string{"first", "second", "third"}
+// ``
 func (a errorMsgs) Errors() []string {
 	if len(a) == 0 {
-		return []string{}
+		return nil
 	}
 	errorStrings := make([]string, len(a))
 	for i, err := range a {