+++ Godeps/_workspace/src/github.com/gin-gonic/gin/render/data.go
@@ -11,9 +11,9 @@ type Data struct {
 	Data        []byte
 }
 
-func (r Data) Write(w http.ResponseWriter) error {
+func (r Data) Render(w http.ResponseWriter) error {
 	if len(r.ContentType) > 0 {
-		w.Header().Set("Content-Type", r.ContentType)
+		w.Header()["Content-Type"] = []string{r.ContentType}
 	}
 	w.Write(r.Data)
 	return nil