+++ Godeps/_workspace/src/github.com/gin-gonic/gin/render/json.go
@@ -19,19 +19,23 @@ type (
 	}
 )
 
-const jsonContentType = "application/json; charset=utf-8"
+var jsonContentType = []string{"application/json; charset=utf-8"}
 
-func (r JSON) Write(w http.ResponseWriter) error {
-	w.Header().Set("Content-Type", jsonContentType)
-	return json.NewEncoder(w).Encode(r.Data)
+func (r JSON) Render(w http.ResponseWriter) error {
+	return WriteJSON(w, r.Data)
 }
 
-func (r IndentedJSON) Write(w http.ResponseWriter) error {
-	w.Header().Set("Content-Type", jsonContentType)
+func (r IndentedJSON) Render(w http.ResponseWriter) error {
+	writeContentType(w, jsonContentType)
 	jsonBytes, err := json.MarshalIndent(r.Data, "", "    ")
 	if err != nil {
 		return err
 	}
 	w.Write(jsonBytes)
 	return nil
+}
+
+func WriteJSON(w http.ResponseWriter, obj interface{}) error {
+	writeContentType(w, jsonContentType)
+	return json.NewEncoder(w).Encode(obj)
 }