+++ Godeps/_workspace/src/github.com/gin-gonic/gin/render/render.go
@@ -7,7 +7,7 @@ package render
 import "net/http"
 
 type Render interface {
-	Write(http.ResponseWriter) error
+	Render(http.ResponseWriter) error
 }
 
 var (
@@ -21,3 +21,10 @@ var (
 	_ HTMLRender = HTMLDebug{}
 	_ HTMLRender = HTMLProduction{}
 )
+
+func writeContentType(w http.ResponseWriter, value []string) {
+	header := w.Header()
+	if val := header["Content-Type"]; len(val) == 0 {
+		header["Content-Type"] = value
+	}
+}