+++ Godeps/_workspace/src/github.com/gin-gonic/gin/render/xml.go
@@ -13,9 +13,9 @@ type XML struct {
 	Data interface{}
 }
 
-const xmlContentType = "application/xml; charset=utf-8"
+var xmlContentType = []string{"application/xml; charset=utf-8"}
 
-func (r XML) Write(w http.ResponseWriter) error {
-	w.Header().Set("Content-Type", xmlContentType)
+func (r XML) Render(w http.ResponseWriter) error {
+	writeContentType(w, xmlContentType)
 	return xml.NewEncoder(w).Encode(r.Data)
 }